core: unión de historias para optimización de documentación en Storybook
