+++ projects/ds-ng/src/lib/components/bmb-list-group/bmb-list-group-item/bmb-list-group-item.component.scss
@@ -1,0 +1,88 @@
+@use "../../../../assets/styles/base/fonts" as fonts;
+@use "../../../../assets/styles/base/rem" as rem;
+
+.bmb_list-group-item {
+  padding: var(--bmb-list-group-item-padding);
+  border: rem.rem_calc(1) solid var(--general_contrasts-container-outline);
+  background-color: var(--general_contrasts-5);
+  display: flex;
+  gap: var(--bmb-spacing-m);
+  align-items: center;
+
+  &-disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+  }
+
+  &-selected {
+    border-color: var(--menu_select-surface-activated);
+    background-color: var(--menu_select-on-surface-hovered);
+  }
+
+  &:hover {
+    &:not(.bmb_list-group-item-selected):not(
+        .bmb_list-group-item-disabled
+      ):hover {
+      border-color: var(--menu_select-surface-activated);
+    }
+  }
+
+  &-content {
+    flex: 1;
+  }
+
+  &-main {
+    display: flex;
+    flex-direction: row;
+
+    @media (max-width: 630px) {
+      flex-direction: column;
+    }
+
+    &-icon {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      padding: 0 var(--bmb-spacing-s);
+
+      @media (max-width: 630px) {
+        justify-content: flex-start;
+      }
+    }
+
+    &-data {
+      display: flex;
+      flex-direction: column;
+      width: 100%;
+      padding: 0 var(--bmb-spacing-s);
+
+      &-header {
+        @include fonts.font(4, medium);
+      }
+
+      &-description {
+        @include fonts.font(4, regular);
+      }
+
+      &-info {
+        @include fonts.font(3, regular);
+      }
+    }
+
+    &-badge {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      padding: 0 var(--bmb-spacing-s);
+
+      @media (max-width: 630px) {
+        justify-content: flex-start;
+        padding: 0 var(--bmb-spacing-m);
+      }
+    }
+  }
+
+  .bmb_field-label {
+    display: none;
+  }
+}
