+++ projects/ds-ng/src/lib/components/bmb-list-group/bmb-list-group.component.ts
@@ -79,82 +79,3 @@ export class BmbListGroupComponent {
     };
   }
 }
-
-@Component({
-  selector: 'bmb-list-group-item',
-  standalone: true,
-  imports: [
-    CommonModule,
-    BmbRadialComponent,
-    BmbCheckboxComponent,
-    BmbIconComponent,
-    BmbTooltipComponent,
-    BmbBadgeComponent,
-    BmbImageComponent,
-  ],
-  templateUrl: './bmb-list-group-item.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  encapsulation: ViewEncapsulation.None,
-})
-export class BmbListGroupItemComponent {
-  id = input.required<string>();
-  isDisabled = input<boolean>(false);
-  isActive = input<boolean>(false);
-  personalizedTemplate = input<boolean>(true);
-  headerText = input<string>('');
-  descriptionText = input<string>('');
-  infoText = input<string>('');
-  icon = input<string>('');
-  imgSrc = input<string>('');
-  tooltipTitle = input<string>('');
-  tooltipText = input<string>('');
-  badgeAppearance = input<IBbmBgAppearance>('mitec_purple');
-  badgeText = input<string>('');
-
-  inputRadioName: string = '';
-  defaultWidthImage = '40px';
-
-  constructor(private bmbListGroupStatusService: BmbListGroupStatusService) {}
-
-  ngOnInit() {
-    if (this.isActive()) {
-      this.bmbListGroupStatusService.updateListGroupStatus(this.id());
-    }
-    this.inputRadioName = this.bmbListGroupStatusService.getListGroupId();
-  }
-
-  handleSelection() {
-    this.bmbListGroupStatusService.updateListGroupStatus(this.id());
-  }
-
-  getClasses(): string[] {
-    const isElementSelected = this.bmbListGroupStatusService
-      .getListGroupStatus()
-      .includes(this.id());
-
-    const classNames = ['bmb_list-group-item'];
-    if (isElementSelected) classNames.push('bmb_list-group-item-selected');
-    if (this.isDisabled()) classNames.push('bmb_list-group-item-disabled');
-
-    return classNames;
-  }
-
-  getConfig() {
-    return this.bmbListGroupStatusService.getListGroupConfiguration();
-  }
-
-  getSelectionLength() {
-    return this.bmbListGroupStatusService.getListGroupStatus().length;
-  }
-
-  showControls(): boolean {
-    return this.bmbListGroupStatusService.getListGroupConfiguration()
-      .showControls;
-  }
-
-  handleItemSelection() {
-    if (!this.showControls() && !this.isDisabled()) {
-      this.bmbListGroupStatusService.updateListGroupStatus(this.id());
-    }
-  }
-}
