+++ projects/ds-ng/src/lib/components/bmb-step-progress-bar/bmb-step-progress-bar.mdx
@@ -1,133 +1,29 @@
-import { Canvas, Meta, Controls } from "@storybook/addon-docs/blocks";
+import {
+  Canvas,
+  Controls,
+  Description,
+  Meta,
+  Title,
+  Subheading,
+} from "@storybook/addon-docs/blocks";
 
 import * as StepProgressBarStories from "./bmb-step-progress-bar.stories";
 import * as StepProgressBarWrapper from "./bmb-step-panel-wrapper.stories";
 
 <Meta of={StepProgressBarStories} />
 
-<div className="sb-container">
-  <div className='sb-section-title'>
-    # Step progress bar
+<Title />
+<Description />
+<Canvas of={StepProgressBarStories.Default} withToolbar={true}/>
 
-    <p>Below is an example of how you can use this component in TypeScript:</p>
-
-  </div>
-</div>
-<div>
-  ```typescript
-  import { BmbStepProgressBarComponent } from '@ti-tecnologico-de-monterrey-oficial/ds-ng';
-  @Component({
-    selector: 'component',
-    standalone: true,
-    imports: [ BmbStepProgressBarComponent ],
-    templateUrl: './component.html',
-    styleUrl: './component.scss',
-  })
-  ```
-</div>
-
-Below is an example of how you can use this component in HTML:
-
-<Canvas of={StepProgressBarStories.Default} />
-<Controls of={StepProgressBarStories.Default} />
-
-## Horizontal step progress bar
-
+<Controls />
+<Subheading>{StepProgressBarStories.Vertical.name}</Subheading>
+<Canvas of={StepProgressBarStories.Vertical} />
+<Subheading>{StepProgressBarStories.Horizontal.name}</Subheading>
 <Canvas of={StepProgressBarStories.Horizontal} />
 
-<div className="sb-container">
-  <div className='sb-section-title'>
-    ## Step progress bar wizard
-
-    <p>Below is an example of how you can use this component in TypeScript:</p>
-
-  </div>
-</div>
-
-```typeScript
-import {
-  Component,
-  ChangeDetectionStrategy,
-  ViewChild,
-  TemplateRef,
-} from '@angular/core';
-import {
-  FormBuilder,
-  FormGroup,
-  Validators,
-  ReactiveFormsModule,
-} from '@angular/forms';
-import { BmbStepProgressBarComponent } from '../../projects/ds-ng/src/public-api';
-import { CommonModule } from '@angular/common';
-
-@Component({
-  selector: 'app-root',
-  standalone: true,
-  imports: [BmbStepProgressBarComponent, CommonModule, ReactiveFormsModule],
-  templateUrl: './app.component.html',
-  styleUrl: './app.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class AppComponent {
-  @ViewChild('step0') step0!: TemplateRef<any>;
-  @ViewChild('step1') step1!: TemplateRef<any>;
-  @ViewChild('step2') step2!: TemplateRef<any>;
-
-  activeStep = 0;
-  stepTemplates: TemplateRef<any>[] = [];
-
-  labelSteps = ['Paso 1', 'Paso 2', 'Paso 3'];
-  forms: FormGroup[] = [];
-
-  constructor(private fb: FormBuilder) {}
-
-  ngOnInit(): void {
-    this.forms = this.labelSteps.map(() =>
-      this.fb.group({
-        campo: ['', Validators.required],
-      }),
-    );
-  }
-
-  ngAfterViewInit(): void {
-    this.stepTemplates = [this.step0, this.step1, this.step2];
-  }
-
-  goNext() {
-    const currentForm = this.forms[this.activeStep];
-    if (currentForm && currentForm.invalid) {
-      currentForm.markAllAsTouched();
-      return;
-    }
-    if (this.activeStep < this.forms.length - 1) {
-      this.activeStep++;
-    }
-  }
-
-  goBack() {
-    if (this.activeStep > 0) {
-      this.activeStep--;
-    }
-  }
-
-  onStepPress(index: number) {
-    const currentForm = this.forms[this.activeStep];
-    if (index > this.activeStep && currentForm && currentForm.invalid) {
-      currentForm.markAllAsTouched();
-      return;
-    }
-    this.activeStep = index;
-  }
-
-  finish() {
-    const currentForm = this.forms[this.activeStep];
-    if (currentForm && currentForm.invalid) {
-      currentForm.markAllAsTouched();
-      return;
-    }
-    alert('¡Formulario finalizado!');
-  }
-}
-```
-
-<Canvas of={StepProgressBarWrapper.StepPanel} />
+<Title of={StepProgressBarWrapper}>{StepProgressBarWrapper.title}</Title>
+<Description of={StepProgressBarWrapper}>
+  {StepProgressBarWrapper.Description}
+</Description>
+<Canvas of={StepProgressBarWrapper.StepPanel} sourceState={'none'} />
